Leaderboard/requirements.txt
Flask==2.2.5

google-cloud-firestore>=2.11.0

gunicorn
